backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.10.3
websockets==14.1
langgraph==0.2.58
langchain==0.3.12
langchain-google-genai==2.0.8
python-multipart==0.0.17
